refactor: feature legacy_ws to enable backwards compatibility
