+++ ethers-providers/src/rpc/transports/legacy_ws.rs
@@ -571,3 +571,31 @@ mod tests {
         resp.unwrap_err();
     }
 }
+
+impl crate::Provider<Ws> {
+    /// Direct connection to a websocket endpoint
+    #[cfg(not(target_arch = "wasm32"))]
+    pub async fn connect(
+        url: impl tokio_tungstenite::tungstenite::client::IntoClientRequest + Unpin,
+    ) -> Result<Self, ProviderError> {
+        let ws = crate::Ws::connect(url).await?;
+        Ok(Self::new(ws))
+    }
+
+    /// Direct connection to a websocket endpoint
+    #[cfg(target_arch = "wasm32")]
+    pub async fn connect(url: &str) -> Result<Self, ProviderError> {
+        let ws = crate::Ws::connect(url).await?;
+        Ok(Self::new(ws))
+    }
+
+    /// Connect to a WS RPC provider with authentication details
+    #[cfg(not(target_arch = "wasm32"))]
+    pub async fn connect_with_auth(
+        url: impl tokio_tungstenite::tungstenite::client::IntoClientRequest + Unpin,
+        auth: Authorization,
+    ) -> Result<Self, ProviderError> {
+        let ws = crate::Ws::connect_with_auth(url, auth).await?;
+        Ok(Self::new(ws))
+    }
+}
