+++ ethers-providers/src/rpc/transports/ws/mod.rs
@@ -20,7 +20,10 @@ use futures_channel::{mpsc, oneshot};
 use serde::{de::DeserializeOwned, Serialize};
 use serde_json::value::RawValue;
 
-use crate::{JsonRpcClient, PubsubClient};
+use crate::{JsonRpcClient, ProviderError, PubsubClient};
+
+#[cfg(not(target_arch = "wasm32"))]
+use crate::Authorization;
 
 #[derive(Debug, Clone)]
 pub struct WsClient {
@@ -96,5 +99,47 @@ impl PubsubClient for WsClient {
         self.instructions
             .unbounded_send(Instruction::Unsubscribe { id: id.into() })
             .map_err(|_| WsClientError::UnexpectedClose)
+    }
+}
+
+impl crate::Provider<WsClient> {
+    /// Direct connection to a websocket endpoint. Defaults to 5 reconnects.
+    pub async fn connect(url: impl Into<ConnectionDetails>) -> Result<Self, ProviderError> {
+        let ws = crate::Ws::connect(url).await?;
+        Ok(Self::new(ws))
+    }
+
+    /// Direct connection to a websocket endpoint, with a set number of
+    /// reconnection attempts
+    pub async fn connect_with_reconnects(
+        url: impl Into<ConnectionDetails>,
+        reconnects: usize,
+    ) -> Result<Self, ProviderError> {
+        let ws = crate::Ws::connect_with_reconnects(url, reconnects).await?;
+        Ok(Self::new(ws))
+    }
+
+    /// Connect to a WS RPC provider with authentication details
+    #[cfg(not(target_arch = "wasm32"))]
+    pub async fn connect_with_auth(
+        url: impl AsRef<str>,
+        auth: Authorization,
+    ) -> Result<Self, ProviderError> {
+        let conn = ConnectionDetails::new(url, Some(auth));
+        let ws = crate::Ws::connect(conn).await?;
+        Ok(Self::new(ws))
+    }
+
+    #[cfg(not(target_arch = "wasm32"))]
+    /// Connect to a WS RPC provider with authentication details and a set
+    /// number of reconnection attempts
+    pub async fn connect_with_auth_and_reconnects(
+        url: impl AsRef<str>,
+        auth: Authorization,
+        reconnects: usize,
+    ) -> Result<Self, ProviderError> {
+        let conn = ConnectionDetails::new(url, Some(auth));
+        let ws = crate::Ws::connect_with_reconnects(conn, reconnects).await?;
+        Ok(Self::new(ws))
     }
 }
